+++ frontend_challenge_2/challenges/tsbs0514/src/mocks/handlers.ts
@@ -1,37 +1,17 @@
 import { http, HttpResponse } from "msw";
-import { PowerArea } from "@/types";
+import { checkAreaLogic } from "@/lib/areaCheck";
 
 /**
  * 郵便番号からエリア判定API
  */
 export const areaCheckHandler = http.post(
-  "/api/area-check",
+  "/e-coding-challenge/api/area-check",
   async ({ request }) => {
     const { postalCode } = (await request.json()) as { postalCode: string };
 
-    const firstDigit = postalCode.charAt(0);
+    const result = checkAreaLogic(postalCode);
 
-    let area: PowerArea = "out-of-service";
-    let isValid = false;
-    let message = "";
-
-    if (firstDigit === "1") {
-      area = "tokyo";
-      isValid = true;
-    } else if (firstDigit === "5") {
-      area = "kansai";
-      isValid = true;
-    } else {
-      area = "out-of-service";
-      isValid = false;
-      message = "サービスエリア対象外です。";
-    }
-
-    return HttpResponse.json({
-      area,
-      isValid,
-      message,
-    });
+    return HttpResponse.json(result);
   }
 );
